pyproject.toml
[tool.poetry]
name = "pricing-engine"
version = "0.0.2"
description = "Pricing Engine"
authors = ["DaRekaze <darekaze@gmail.com>"]
readme = "README.md"
packages = [
    { include = "pricing_engine"},
]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.20.0"
aiometer = "^0.3.0"
httpx = {extras = ["http2", "socks"], version = "^0.23.0"}
websockets = "^10.3"

[tool.poetry.dev-dependencies]
pytest = "^7.1"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
